+++ pagamento.cpp
@@ -2,7 +2,7 @@
 
 double Pagamento::_pesoPagamento=30; // => l'incarico vale 1 punto ogni 30 euro di importo
 
-Pagamento::Pagamento(const std::string& nome, double importo, Inquilino *incaricato, bool svolto):
+Pagamento::Pagamento(const std::string& nome, float importo, Inquilino *incaricato, bool svolto):
     Incarico(nome,incaricato,svolto),_importo(importo) {}
 
 void Pagamento::setImporto(double x){
